analyzer/evm: move ERC-20 to its own file
